Added Auth

+++ app/api/v1/apartments/route.tsx
@@ -1,19 +1,28 @@
-// app/api/apartments/route.tsx
 import { NextRequest, NextResponse } from 'next/server';
 import { PrismaClient, Apartment } from '@prisma/client';
+import { authenticate } from '../auth/auth';
 
 const prisma = new PrismaClient();
+const allowedRoles = ['admin', 'user'];
 
 /**
  * @swagger
  * /apartments:
  *   get:
  *     summary: Retrieve a list of apartments
+ *     security:
+ *       - bearerAuth: []
  *     responses:
  *       200:
  *         description: List of apartments
+ *       401:
+ *         description: Unauthorized
+ *       403:
+ *         description: Forbidden: Access Denied
  *   post:
  *     summary: Create a new apartment
+ *     security:
+ *       - bearerAuth: []
  *     requestBody:
  *       required: true
  *       content:
@@ -37,6 +46,10 @@ const prisma = new PrismaClient();
  *     responses:
  *       201:
  *         description: Apartment created
+ *       401:
+ *         description: Unauthorized
+ *       403:
+ *         description: Forbidden: Access Denied
  *   put:
  *     summary: Update an existing apartment
  *     requestBody:
@@ -59,6 +72,10 @@ const prisma = new PrismaClient();
  *     responses:
  *       200:
  *         description: Apartment updated
+ *       401:
+ *         description: Unauthorized
+ *       403:
+ *         description: Forbidden: Access Denied
  *   delete:
  *     summary: Delete an apartment
  *     parameters:
@@ -71,14 +88,25 @@ const prisma = new PrismaClient();
  *     responses:
  *       200:
  *         description: Apartment deleted
+ *       401:
+ *         description: Unauthorized
+ *       403:
+ *         description: Forbidden: Access Denied
  */
 
 export async function GET(request: NextRequest) {
+  // const authResult = await authenticate(request, allowedRoles);
+  const token = await authenticate(request);
+  if (token !== null) return token;
+  
   const apartments = await prisma.apartment.findMany();
   return NextResponse.json(apartments);
 }
 
 export async function POST(request: NextRequest) {
+  const token = await authenticate(request);
+  if (token !== null) return token;
+  
   const data = await request.json();
   const newApartment = await prisma.apartment.create({
     data,
@@ -87,6 +115,9 @@ export async function POST(request: NextRequest) {
 }
 
 export async function PUT(request: NextRequest) {
+  const token = await authenticate(request);
+  if (token !== null) return token;
+  
   const { id, ...data } = await request.json();
   const updatedApartment = await prisma.apartment.update({
     where: { id },
@@ -96,6 +127,9 @@ export async function PUT(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
+  const token = await authenticate(request);
+  if (token !== null) return token;
+  
   const { searchParams } = new URL(request.url);
   const id = parseInt(searchParams.get('id') || '');
   const deletedApartment = await prisma.apartment.delete({
